test: split oversized public contract suites
