+++ packages/ui/src/product/converter/advanced-mode/sections/proxy-groups-added-rule-sets.test.ts
@@ -673,58 +673,4 @@ describe("ProxyGroupsAddedRuleSets", () => {
     expect(stateMock.setters[1]).toHaveBeenCalledWith(null);
   });
 
-  it("ignores invalid saves and reports missing targets as conflicts", () => {
-    renderAdded({
-      0: moduleItem.key,
-      1: { path: "geosite/rule-a.mrs", targetValue: "bad", noResolve: false },
-    });
-    mocks.captures.buttons
-      .find((props: any) => props.title === "保存规则集")
-      .onClick();
-    expect(mocks.store.updateModuleRule).not.toHaveBeenCalled();
-
-    renderAdded({
-      0: moduleItem.key,
-      1: {
-        path: "geosite/rule-a.mrs",
-        targetValue: "module:missing",
-        noResolve: false,
-      },
-    });
-    mocks.captures.buttons
-      .find((props: any) => props.title === "保存规则集")
-      .onClick();
-    expect(mocks.toast).toHaveBeenCalledWith(
-      expect.objectContaining({ title: "规则集已存在", variant: "warning" }),
-    );
-
-    renderAdded({
-      0: customItem.key,
-      1: {
-        path: "geoip/rule-b.mrs",
-        targetValue: "custom:missing",
-        noResolve: false,
-      },
-    });
-    mocks.captures.buttons
-      .find((props: any) => props.title === "保存规则集")
-      .onClick();
-    expect(mocks.toast).toHaveBeenCalledWith(
-      expect.objectContaining({ title: "规则集已存在", variant: "warning" }),
-    );
-
-    renderAdded({
-      0: moduleItem.key,
-      1: { path: "   ", targetValue: "module:auto", noResolve: false },
-    });
-    mocks.captures.buttons
-      .find((props: any) => props.title === "保存规则集")
-      .onClick();
-    expect(mocks.store.updateModuleRule).not.toHaveBeenCalledWith(
-      "auto",
-      "rule-a",
-      expect.objectContaining({ path: "" }),
-    );
-
-  });
 });
